增加页面附件和意见组件的生成。

+++ core/mybatis-generator-core/src/main/java/org/mybatis/generator/api/MyBatisGenerator.java
@@ -337,9 +337,11 @@ public class MyBatisGenerator {
             }
 
             callback.checkCancel();
-            callback.startTask(getString(
-                    "Progress.15", targetFile.getName())); //$NON-NLS-1$
-            writeFile(targetFile, source, gf.getFileEncoding());
+            if (!targetFile.exists() || gf.isOverWriteFile()) {
+                callback.startTask(getString(
+                        "Progress.15", targetFile.getName())); //$NON-NLS-1$
+                writeFile(targetFile, source, gf.getFileEncoding());
+            }
         } catch (ShellException e) {
             warnings.add(e.getMessage());
         }
